new layout addedd to ToDoList web program
