+++ Pages/Index.cshtml.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -11,7 +10,7 @@ namespace redbus.Pages
 {
     public class IndexModel : PageModel
     {
-        public string[] allTask;
+        
         private readonly ILogger<IndexModel> _logger;
 
         public IndexModel(ILogger<IndexModel> logger)
@@ -21,14 +20,8 @@ namespace redbus.Pages
 
         public void OnGet()
         {
-            allTask=System.IO.File.ReadAllLines("dotasks.txt");
+          
         }
-        public void OnPost()
-        {
-           string task=Request.Form["dolist"];
-           System.IO.File.AppendAllText("dotasks.txt",task+Environment.NewLine);
-           allTask=System.IO.File.ReadAllLines("dotasks.txt");
-           
-        }
+      
     }
 }
