+++ dotasks.txt
@@ -1,2 +1,13 @@
-dotasks
+drink water
 read texts
+asdsafdsf
+do tasks
+complete works
+drink water
+jog
+run
+walk
+hike
+do tasks
+drink water
+asddf
